+++ include/crm/common/internal.h
@@ -13,6 +13,7 @@
 #include <glib.h>       /* for gboolean */
 #include <dirent.h>     /* for struct dirent */
 #include <unistd.h>     /* for getpid() */
+#include <stdbool.h>    /* for bool */
 #include <sys/types.h>  /* for uid_t and gid_t */
 
 #include <crm/common/logging.h>
@@ -32,6 +33,8 @@ char *crm_read_contents(const char *filename);
 int crm_write_sync(int fd, const char *contents);
 int crm_set_nonblocking(int fd);
 const char *crm_get_tmpdir(void);
+
+void pcmk__close_fds_in_child(bool);
 
 
 /* internal procfs utilities (from procfs.c) */
